Moved rest service callbacks init to separate file
