+++ IOTbase/platformio.ini
@@ -6,15 +6,6 @@ lib_deps = NeoPixelBus
            WebSockets
 lib_extra_dirs = ../lib
 
-[env:esp01]
-extra_script = extra_script.py
-platform = espressif8266
-board = esp01
-build_flags = ${common.build_flags}
-framework = arduino
-lib_deps = ${common.lib_deps}
-lib_extra_dirs = ${common.lib_extra_dirs}
-
 [env:d1_mini]
 extra_script = extra_script.py
 platform = espressif8266
@@ -26,5 +17,5 @@ framework = arduino
 lib_deps = ${common.lib_deps}
 lib_extra_dirs = ${common.lib_extra_dirs}
 
-#upload_port = 192.168.4.1
-#upload_flags = --port=8266 --auth=admin
+upload_port = 192.168.4.1
+upload_flags = --port=8266 --auth=admin
